requirements.txt
motorcortex-python==0.20.14
opencv-python==4.5.2.54
numpy==1.19.5
pypot==3.1.3
